Move enforcer to external service
