+++ core/src/service/ProjectBoardService.ts
@@ -1,4 +1,6 @@
 import { EventEmitter } from 'events';
+import { Actions, EnforcerService, PolicyApplicationFailedError } from '../../../external/enforcer/src/EnforcerService';
+import { Resources } from '../../../external/enforcer/src/resources/resources';
 import { CreateBoardEntryQuery } from '../access/queries/CreateBoardEntryQuery';
 import { DeleteBoardEntryQuery } from '../access/queries/DeleteBoardEntryQuery';
 import { DescribeBoardEntryQuery } from '../access/queries/DescribeBoardEntryQuery';
@@ -8,8 +10,6 @@ import { BoardEntryImmutable } from '../schemas/types/ProjectBoard/BoardEntryImm
 import { BoardEntryMaster } from '../schemas/types/ProjectBoard/BoardEntryMaster';
 import { BoardEntryMutable } from '../schemas/types/ProjectBoard/BoardEntryMutable';
 import { TranslateErrors } from './decorators';
-import { Actions, EnforcerService, PolicyApplicationFailedError } from './enforcer/EnforcerService';
-import { Resources } from './enforcer/resources/resources';
 import * as Utils from './util';
 
 type CreateBoardEntryParams = Utils.AuthenticatedSingleResourceServiceCall<BoardEntryImmutable>;
