+++ external/enforcer/src/policies/InvitePolicy.ts
@@ -1,6 +1,6 @@
-import { Claims } from "../../../access/auth/verify";
-import InviteModel from "../../../access/models/InviteModel";
-import { describeMembership } from "../../../access/queries/util";
+import { Claims } from "../../../../core/src/access/auth/verify";
+import InviteModel from "../../../../core/src/access/models/InviteModel";
+import { describeMembership } from "../../../../core/src/access/queries/util";
 import { Actions, Policy } from "../EnforcerService";
 import { Resources } from "../resources/resources";
