+++ external/enforcer/src/resources/resources.ts
