+++ tsconfig.json
@@ -19,7 +19,8 @@
 	},
     "include": [
         "core/src/**/*",
-        "external/**/src/**/*"
+        "external/**/src/**/*", 
+        "external/enforcer"
     ],
     
 	// keep build steps out of deployment
